+++ packages/common/src/index.ts
@@ -3,7 +3,6 @@ export { BucketsConfig };
 
 export * as cliUtils from './cli-utils';
 export * as http from './http';
-export * as keyPairProvider from './key-pair-provider';
 export * as log from './log';
 export * as testUtils from './test-utils';
 export * as util from './util';